requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
cffi==1.11.5
cryptography==2.3.1
cx-Oracle==7.0.0
Django==1.11.15
django-background-tasks==1.1.13
django-bootstrap3==11.0.0
django-compat==1.0.15
idna==2.7
paramiko==2.4.1
pyasn1==0.4.4
pycparser==2.18
pycryptodome==3.6.6
PyMySQL==0.9.2
PyNaCl==1.2.1
pytz==2018.5
six==1.11.0
sshtunnel==0.1.4
